remove redundant cast

+++ src/main/java/water/fvec/C1SChunk.java
@@ -54,7 +54,7 @@ public class C1SChunk extends Chunk {
     for( int i=0; i<len; i++ ) {
       int res = 0xFF&_mem[i+OFF];
       if( res == C1Chunk._NA ) nc.addNA();
-      else nc.addNum((long)(res+_bias),(int)dx);
+      else nc.addNum((res+_bias),(int)dx);
     }
     return nc;
   }
